Fixed a warning on 64 bit architectures.

+++ drbd/drbd_req.c
@@ -1100,7 +1100,8 @@ int drbd_make_request_26(struct request_queue *q, struct bio *bio)
 		/* rather error out here than BUG in bio_split */
 		ERR("bio would need to, but cannot, be split: "
 		    "(vcnt=%u,idx=%u,size=%u,sector=%llu)\n",
-		    bio->bi_vcnt, bio->bi_idx, bio->bi_size, bio->bi_sector);
+		    bio->bi_vcnt, bio->bi_idx, bio->bi_size, 
+		    (unsigned long long)bio->bi_sector);
 		bio_endio(bio, -EINVAL);
 		return 0;
 	} else {
